Separate rules from data.policies definition and evaluation
Now, helm templates generate one ConfigMap for the OPA rules, and other for the Policy settings from values.yaml as a policy.json file. This policy.json is imported as data separately. This makes it easier to change configuration once deployed.

The scanRules section in values.yaml is now renamed to scanPolicies.

Other minor changes:
* Fix nil reference if loading OPA rules or data failed.
* Add option to specify image tag in values.yaml.
* Add definition of preScanPolicies and preCustomRules, for evaluating rules before triggering image scan (TO-DO).
* Upgrade Go Builder to 1.14.2.

Signed-off-by: Alvaro Iradier <alvaro.iradier@sysdig.com>